refactor: extract debate engine from project.js into project-debate.js
PR-01 of the refactoring roadmap. Moves all 27 debate functions
(~1,500 lines) out of the 7,222-line project.js god object into
a dedicated project-debate.js module. Uses the context object
pattern so the new module receives only the state and callbacks
it needs, with no circular imports. All call sites in project.js
remain unchanged via thin re-export variables. Also adds the
granular REFACTORING_ROADMAP.md (42 PRs) for coding agents.
